lab/knn/knn_surgery/requirements.txt
Pillow
coremltools